build: exclude vue components from api report (refs SFKUI-6500)
