+++ packages/vue-labs/src/api.ts
@@ -1,0 +1,1 @@
+export {};
