+++ packages/vue-labs/src/index.ts
@@ -1,1 +1,5 @@
+/* include typescript declarations for vite static asset handling (e.g. `?raw`) */
+/// <reference types="vite/client" />
+
+export * from "./api";
 export * from "./components";
