+++ packages/vue/api-extractor.api.json
@@ -1,20 +1,13 @@
 {
     "$schema": "https://developer.microsoft.com/json-schemas/api-extractor/v7/api-extractor.schema.json",
     "extends": "../../api-extractor.json",
-    "apiReport": {
-        "enabled": false
-    },
+    "mainEntryPointFilePath": "<projectFolder>/temp/types/api.d.ts",
     "dtsRollup": {
-        "publicTrimmedFilePath": "<projectFolder>/dist/types/index.d.ts"
+        "enabled": false
     },
     "messages": {
         "compilerMessageReporting": {
             "default": {
-                "logLevel": "none"
-            }
-        },
-        "extractorMessageReporting": {
-            "ae-forgotten-export": {
                 "logLevel": "none"
             }
         }
