+++ packages/vue/api-extractor.lib.json
@@ -1,12 +1,20 @@
 {
     "$schema": "https://developer.microsoft.com/json-schemas/api-extractor/v7/api-extractor.schema.json",
     "extends": "../../api-extractor.json",
+    "apiReport": {
+        "enabled": false
+    },
     "dtsRollup": {
         "publicTrimmedFilePath": "<projectFolder>/dist/types/index.d.ts"
     },
     "messages": {
         "compilerMessageReporting": {
             "default": {
+                "logLevel": "none"
+            }
+        },
+        "extractorMessageReporting": {
+            "ae-forgotten-export": {
                 "logLevel": "none"
             }
         }
