+++ packages/vue/src/api.ts
@@ -1,0 +1,30 @@
+export {
+    type UseSlotUtils,
+    type WithoutInstance,
+    useSlotUtils,
+} from "./composables";
+export * from "./components/api";
+export {
+    config,
+    type FKUIConfig,
+    FKUIConfigButtonOrder,
+    type FKUIContext,
+    type MaybeWithFKUIContext,
+    setRunningContext,
+} from "./config";
+export * from "./internal-components/api";
+export * from "./plugins";
+export {
+    type ComponentUnmountEvent,
+    type ComponentValidityEvent,
+    type ErrorItem,
+    type GroupValidityEvent,
+    type ListArray,
+    type ListItem,
+    type UnknownItem,
+    ErrorData,
+    ErrorViewData,
+    FormErrorList,
+    MenuAction,
+} from "./types";
+export * from "./utils";
