+++ packages/vue/src/components/index.ts
@@ -1,130 +1,50 @@
+/* component exports only, non-components go in api.ts */
+
+export * from "./api";
+
 export { FBadge } from "./FBadge";
 export { FCalendar, FCalendarDay } from "./FCalendar";
 export { FCard } from "./FCard";
 export { FCheckboxField } from "./FCheckboxField";
-export {
-    type ContextMenuItem,
-    type ContextMenuSeparatorItem,
-    type ContextMenuTextItem,
-    FContextMenu,
-    isContextMenuSeparatorItem,
-    isContextMenuTextItem,
-} from "./FContextMenu";
-export {
-    type ActivateItemCallback,
-    type ActivateItemInterface,
-    type FCrudDatasetInterface,
-    ActivateItemInjected,
-    FCrudButton,
-    FCrudDataset,
-    Operation,
-} from "./FCrudDataset";
+export { FConfirmModal, FFormModal, FModal } from "./FModal";
+export { FContextMenu } from "./FContextMenu";
+export { FCrudButton, FCrudDataset } from "./FCrudDataset";
 export { FDataTable } from "./FDataTable";
 export { FDatepickerField } from "./FDatepickerField";
-export {
-    type DetailsPanelCloseCallback,
-    type UseDetailsPanel,
-    FDetailsPanel,
-    useDetailsPanel,
-} from "./FDetailsPanel";
-export {
-    type FDialogueTreeData,
-    type FDialogueTreeEndQuestion,
-    type FDialogueTreeOption,
-    type FDialogueTreeQuestion,
-    type FDialogueTreeSubQuestion,
-    type FDialogueTreeUserProgress,
-    FDialogueTree,
-    isDialogueTreeEndQuestion,
-} from "./FDialogueTree";
-export { type BeforeNavigate, FErrorList } from "./FErrorList";
+export { FDetailsPanel } from "./FDetailsPanel";
+export { FDialogueTree } from "./FDialogueTree";
+export { FErrorList } from "./FErrorList";
 export { FExpand } from "./FExpand";
 export { FExpandablePanel } from "./FExpandablePanel";
 export { FExpandableParagraph } from "./FExpandableParagraph";
 export { FFieldset } from "./FFieldset";
-export { type FFileItemIconName, FFileItem } from "./FFileItem";
+export { FFileItem } from "./FFileItem";
 export { FFileSelector } from "./FFileSelector";
+export { FFixedPane } from "./FFixedPane";
 export { FIcon } from "./FIcon";
 export { FInteractiveTable } from "./FInteractiveTable";
-export { FFixedPane } from "./FFixedPane";
 export { FLabel } from "./FLabel";
 export { FLayoutApplicationTemplate } from "./FLayoutApplicationTemplate";
-export {
-    type PanelLayoutComposable,
-    FLayoutLeftPanel,
-} from "./FLayoutLeftPanel";
-export {
-    type FLayoutRightPanelInteface,
-    type FLayoutRightPanelContentEvent,
-    FLayoutRightPanel,
-    FLayoutRightPanelService,
-} from "./FLayoutRightPanel";
+export { FLayoutLeftPanel } from "./FLayoutLeftPanel";
+export { FLayoutRightPanel } from "./FLayoutRightPanel";
 export { FList } from "./FList";
 export { FLoader } from "./FLoader";
 export { FLogo } from "./FLogo";
 export { FMessageBox } from "./FMessageBox";
-export {
-    type FModalButton,
-    type FModalButtonDescriptor,
-    type FModalData,
-    FConfirmModal,
-    FFormModal,
-    FFormModalAction,
-    FModal,
-} from "./FModal";
-export {
-    type FNavigationMenuData,
-    type MenuItem,
-    type NavigationMenuItem,
-    FNavigationMenu,
-} from "./FNavigationMenu";
+export { FMinimizablePanel } from "./FMinimizablePanel";
+export { FNavigationMenu } from "./FNavigationMenu";
 export { FOffline } from "./FOffline";
 export { FOutputField } from "./FOutputField";
 export { FPageHeader } from "./FPageHeader";
-export {
-    type FPageLayoutBindings,
-    type FPageLayoutType,
-    type FPageLayoutSlotMapping,
-    type LayoutDefinition,
-    type LayoutAreaAttachPanel,
-    type LayoutAreaDefinition,
-    type LayoutAreaDirection,
-    type UseAreaData,
-    FPageLayout,
-    registerLayout,
-    useAreaData,
-} from "./FPageLayout";
-export { FMinimizablePanel } from "./FMinimizablePanel";
+export { FPageLayout } from "./FPageLayout";
 export { FProgressbar } from "./FProgressbar";
 export { FRadioField } from "./FRadioField";
-export {
-    type UseResize,
-    type UseResizeOptions,
-    FResizePane,
-    useResize,
-} from "./FResizePane";
+export { FResizePane } from "./FResizePane";
 export { FSelectField } from "./FSelectField";
-export {
-    type FSortFilterDatasetInterface,
-    type FSortFilterDatasetMountCallback,
-    type FSortFilterDatasetSortCallback,
-    type SortOrder,
-    FSortFilterDataset,
-    FSortFilterDatasetInjected,
-} from "./FSortFilterDataset";
+export { FSortFilterDataset } from "./FSortFilterDataset";
 export { FStaticField } from "./FStaticField";
+export { FTableColumn } from "./FTableColumn";
 export {
-    type FTableColumnData,
-    type FTableInterface,
-    FTableColumn,
-    FTableColumnSize,
-    FTableColumnSort,
-    FTableColumnType,
-} from "./FTableColumn";
-export {
-    type FormatFunction,
-    type ParseFunction,
-    type TextFieldSetupProps,
     FBankAccountNumberTextField,
     FBankgiroTextField,
     FClearingnumberTextField,
@@ -139,26 +59,9 @@ export {
     FPostalCodeTextField,
     FSearchTextField,
     FTextField,
-    useTextFieldSetup,
 } from "./FTextField";
 export { FTextareaField } from "./FTextareaField";
 export { FTooltip, tooltipAttachTo } from "./FTooltip";
-export {
-    type FValidationFormCallback,
-    type FValidationFormResult,
-    FValidationForm,
-    FValidationFormAction,
-    IValidationForm,
-} from "./FValidationForm";
+export { FValidationForm, IValidationForm } from "./FValidationForm";
 export { FValidationGroup } from "./FValidationGroup";
-export {
-    type FWizardApi,
-    type FWizardStepDefinition,
-    type FWizardKey,
-    type StepNumber,
-    type FWizardValidationData,
-    type FWizardValidationCallback,
-    FWizard,
-    FWizardStep,
-    FWizardStepAction,
-} from "./FWizard";
+export { FWizard, FWizardStep } from "./FWizard";
