+++ packages/vue/src/internal-components/api.ts
@@ -1,0 +1,7 @@
+/* non-component exports only, components go in index.ts */
+
+export { type IPopupData, Placement } from "./IPopup";
+export { type IPopupErrorData } from "./IPopupError";
+export { type AnimationCallback } from "./IAnimateExpand";
+export { getDayEndOffset, getDayStartOffset } from "./calendar";
+export { useCombobox } from "./combobox";
