+++ packages/vue/src/internal-components/index.ts
@@ -1,20 +1,17 @@
-export { IFlex, IFlexItem } from "./IFlex";
-export { type IPopupData, IPopup, Placement } from "./IPopup";
-export { type IPopupErrorData, IPopupError } from "./IPopupError";
-export { IPopupListbox } from "./IPopupListbox";
-export { IPopupMenu } from "./IPopupMenu";
-export { type AnimationCallback, IAnimateExpand } from "./IAnimateExpand";
-export { ISkipLink } from "./ISkipLink";
+/* component exports only, non-components go in api.ts */
+
+export * from "./api"
+
 export {
-    getDayEndOffset,
-    getDayStartOffset,
     ICalendarMonth,
     ICalendarMonthGrid,
     ICalendarNavbar,
 } from "./calendar";
-
-export {
-    IComboboxDropdown,
-    IComboboxToggleButton,
-    useCombobox,
-} from "./combobox";
+export { IComboboxDropdown, IComboboxToggleButton } from "./combobox";
+export { IAnimateExpand } from "./IAnimateExpand";
+export { IFlex, IFlexItem } from "./IFlex";
+export { IPopup } from "./IPopup";
+export { IPopupError } from "./IPopupError";
+export { IPopupListbox } from "./IPopupListbox";
+export { IPopupMenu } from "./IPopupMenu";
+export { ISkipLink } from "./ISkipLink";
